Updated, Roles, Esports, message_handel.py, for enhancement

+++ src/modules/message_handel.py
@@ -76,7 +76,6 @@ async def tourney(message):
     guild = message.guild
     td = tourneydbc.find_one({"tid" : message.channel.id%1000000000000})
     tmrole = discord.utils.get(ctx.guild.roles, name="tourney-mod")
-
     if message.author.bot:
         return
       
@@ -98,10 +97,19 @@ async def tourney(message):
         rgs = td["reged"]
         tslot = td["tslot"]
 
+        if not crole:
+            try:
+                await message.author.send("Registration Paused")
+            except:
+                pass
+            await ctx.reply("Confirm Role Not Found")
+            return dbc.update_one({"tid" : td["rch"]}, {"$set" : {"status" : "paused"}})
+
+
             
         if crole in message.author.roles:
             await message.delete()
-            return await message.channel.send("Already Registered", delete_after=5)
+            return await message.channel.send("**Already Registered**", delete_after=5)
             
 
         if rgs > tslot:
@@ -115,6 +123,11 @@ async def tourney(message):
             
             for fmsg in messages:
                 tk = len(set(ctx.mentions) & set(fmsg.mentions))
+
+
+#IF FAKE TAG NOT ALLOWED
+########################
+
                 if td["faketag"] == "no":
 
  
@@ -124,7 +137,7 @@ async def tourney(message):
                             await message.add_reaction("✅")
                             reg_update(message)
                             team_name = find_team(message)
-                            femb = discord.Embed(color=0xffff00, description=f"**{rgs}) TEAM NAME: [{team_name.upper()}](https://discordapp.com/channels/{message.channel.guild.id}/{message.channel.id}/{message.id})**\n**Players** : {(', '.join(m.mention for m in message.mentions)) if message.mentions else message.author.mention} ")
+                            femb = discord.Embed(color=0xffff00, description=f"**{rgs}) TEAM NAME: [{team_name.upper()}]({message.jump_url})**\n**Players** : {(', '.join(m.mention for m in message.mentions)) if message.mentions else message.author.mention} ")
                             femb.set_author(name=message.guild.name, icon_url=message.guild.icon)
                             femb.timestamp = datetime.datetime.utcnow()
                             await cch.send(f"{team_name.upper()} {message.author.mention}", embed=femb)
@@ -133,17 +146,14 @@ async def tourney(message):
                                 dbc.update_one({"rch" : rch.id}, {"$set" : {"pub" : "yes", "prize" : "Not Published Yet"}})
 
 
-
                     if fmsg.author.id != ctx.author.id:
                         ftch = await ft_ch(message)
                         if ftch == True:
                             fakeemb = discord.Embed(title=f"The Member You Tagged is Already Registered In A Team. If You Think He Used `Fake Tags`, You can Contact `Management Team`", color=0xffff00)
-                            fakeemb.add_field(name="Team", value=f"[Registration Link](https://discordapp.com/channels/{guild.id}/{message.channel.id}/{fmsg.id})")
+                            fakeemb.add_field(name="Team", value=f"[Registration Link]({fmsg.jump_url})")
                             fakeemb.set_author(name=ctx.author, icon_url=ctx.author.avatar)
                             await message.delete()
                             return await ctx.channel.send(embed=fakeemb, delete_after=60)
-
-                            
                                 
 
                         if ftch != True:
@@ -151,20 +161,20 @@ async def tourney(message):
                             await message.add_reaction("✅")
                             reg_update(message)
                             team_name = find_team(message)
-                            femb = discord.Embed(color=0xffff00, description=f"**{rgs}) TEAM NAME: [{team_name.upper()}](https://discordapp.com/channels/{message.channel.guild.id}/{message.channel.id}/{message.id})**\n**Players** : {(', '.join(m.mention for m in message.mentions)) if message.mentions else message.author.mention} ")
+                            femb = discord.Embed(color=0xffff00, description=f"**{rgs}) TEAM NAME: [{team_name.upper()}]({message.jump_url})**\n**Players** : {(', '.join(m.mention for m in message.mentions)) if message.mentions else message.author.mention} ")
                             femb.set_author(name=message.guild.name, icon_url=message.guild.icon)
                             femb.timestamp = datetime.datetime.utcnow()
                             return await cch.send(f"{team_name.upper()} {message.author.mention}", embed=femb)
                         
-
-                            
+#IF FAKE TAG ALLOWED
+####################
 
                 if td["faketag"] == "yes":
                     await message.author.add_roles(crole)
                     await message.add_reaction("✅")
                     reg_update(message)
                     team_name = find_team(message)
-                    nfemb = discord.Embed(color=0xffff00, description=f"**{rgs}) TEAM NAME: [{team_name.upper()}](https://discordapp.com/channels/{message.channel.guild.id}/{message.channel.id}/{message.id})**\n**Players** : {(', '.join(m.mention for m in message.mentions)) if message.mentions else message.author.mention} ")
+                    nfemb = discord.Embed(color=0xffff00, description=f"**{rgs}) TEAM NAME: [{team_name.upper()}]({message.jump_url})**\n**Players** : {(', '.join(m.mention for m in message.mentions)) if message.mentions else message.author.mention} ")
                     nfemb.set_author(name=message.guild.name, icon_url=message.guild.icon)
                     nfemb.timestamp = datetime.datetime.utcnow()
                     return await cch.send(f"{team_name.upper()} {message.author.mention}", embed=nfemb)
@@ -172,7 +182,7 @@ async def tourney(message):
 
         elif len(message.mentions) < ments:
             #await bot.process_commands(message)
-            meb = discord.Embed(description=f"Minimum {ments} Mentions Required For Successfull Registration", color=0xff0000)
+            meb = discord.Embed(description=f"**Minimum {ments} Mentions Required For Successfull Registration**", color=0xff0000)
             await message.delete()
             return await message.channel.send(content=message.author.mention, embed=meb, delete_after=5)
